requirements.txt
mediapipe>=0.10.9
opencv-python>=4.8.0
pynput>=1.7.6
pystray>=0.19.5
flask>=3.0.0
Pillow>=10.0.0
pytest>=7.0
pytest-mock>=3.0
